refactoring of home and forms controllers. Removed controller forms.
